// src/utils/server/backend/model/pipeline-inspection-value.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface PipelineInspectionValue
 */
export interface PipelineInspectionValue {
  /**
   *
   * @type {number}
   * @memberof PipelineInspectionValue
   */
  value: number | null;
  /**
   *
   * @type {string}
   * @memberof PipelineInspectionValue
   */
  unit: string | null;
}